#0: Cleanup permute: split hpp and cpp
